Abstract away entry procedure in tests
